// src/api/generated/models/GroupResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eGovFrame 5.0 API
 * API Documentation for eGovFrame 5.0 Application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GroupResponse
 */
export interface GroupResponse {
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    groupId?: string;
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    groupName?: string;
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    groupCode?: string;
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    groupDescription?: string;
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    branchId?: string;
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    branchName?: string;
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    useAt?: string;
    /**
     * 
     * @type {Date}
     * @memberof GroupResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof GroupResponse
     */
    updatedDate?: Date;
}

/**
 * Check if a given object implements the GroupResponse interface.
 */
export function instanceOfGroupResponse(value: object): value is GroupResponse {
    return true;
}

export function GroupResponseFromJSON(json: any): GroupResponse {
    return GroupResponseFromJSONTyped(json, false);
}

export function GroupResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GroupResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'groupId': json['groupId'] == null ? undefined : json['groupId'],
        'groupName': json['groupName'] == null ? undefined : json['groupName'],
        'groupCode': json['groupCode'] == null ? undefined : json['groupCode'],
        'groupDescription': json['groupDescription'] == null ? undefined : json['groupDescription'],
        'branchId': json['branchId'] == null ? undefined : json['branchId'],
        'branchName': json['branchName'] == null ? undefined : json['branchName'],
        'tenantId': json['tenantId'] == null ? undefined : json['tenantId'],
        'useAt': json['useAt'] == null ? undefined : json['useAt'],
        'createdDate': json['createdDate'] == null ? undefined : (new Date(json['createdDate'])),
        'updatedDate': json['updatedDate'] == null ? undefined : (new Date(json['updatedDate'])),
    };
}

export function GroupResponseToJSON(json: any): GroupResponse {
    return GroupResponseToJSONTyped(json, false);
}

export function GroupResponseToJSONTyped(value?: GroupResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'groupId': value['groupId'],
        'groupName': value['groupName'],
        'groupCode': value['groupCode'],
        'groupDescription': value['groupDescription'],
        'branchId': value['branchId'],
        'branchName': value['branchName'],
        'tenantId': value['tenantId'],
        'useAt': value['useAt'],
        'createdDate': value['createdDate'] == null ? undefined : ((value['createdDate']).toISOString()),
        'updatedDate': value['updatedDate'] == null ? undefined : ((value['updatedDate']).toISOString()),
    };
}