// src/api/generated/models/BranchResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eGovFrame 5.0 API
 * API Documentation for eGovFrame 5.0 Application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BranchResponse
 */
export interface BranchResponse {
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    branchId?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    branchName?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    branchCode?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    branchAddress?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    branchPhone?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    parentBranchId?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    useAt?: string;
    /**
     * 
     * @type {Date}
     * @memberof BranchResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof BranchResponse
     */
    updatedDate?: Date;
}

/**
 * Check if a given object implements the BranchResponse interface.
 */
export function instanceOfBranchResponse(value: object): value is BranchResponse {
    return true;
}

export function BranchResponseFromJSON(json: any): BranchResponse {
    return BranchResponseFromJSONTyped(json, false);
}

export function BranchResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): BranchResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'branchId': json['branchId'] == null ? undefined : json['branchId'],
        'branchName': json['branchName'] == null ? undefined : json['branchName'],
        'branchCode': json['branchCode'] == null ? undefined : json['branchCode'],
        'branchAddress': json['branchAddress'] == null ? undefined : json['branchAddress'],
        'branchPhone': json['branchPhone'] == null ? undefined : json['branchPhone'],
        'parentBranchId': json['parentBranchId'] == null ? undefined : json['parentBranchId'],
        'tenantId': json['tenantId'] == null ? undefined : json['tenantId'],
        'useAt': json['useAt'] == null ? undefined : json['useAt'],
        'createdDate': json['createdDate'] == null ? undefined : (new Date(json['createdDate'])),
        'updatedDate': json['updatedDate'] == null ? undefined : (new Date(json['updatedDate'])),
    };
}

export function BranchResponseToJSON(json: any): BranchResponse {
    return BranchResponseToJSONTyped(json, false);
}

export function BranchResponseToJSONTyped(value?: BranchResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'branchId': value['branchId'],
        'branchName': value['branchName'],
        'branchCode': value['branchCode'],
        'branchAddress': value['branchAddress'],
        'branchPhone': value['branchPhone'],
        'parentBranchId': value['parentBranchId'],
        'tenantId': value['tenantId'],
        'useAt': value['useAt'],
        'createdDate': value['createdDate'] == null ? undefined : ((value['createdDate']).toISOString()),
        'updatedDate': value['updatedDate'] == null ? undefined : ((value['updatedDate']).toISOString()),
    };
}